+++ janitor/__init__.py
@@ -1,10 +1,10 @@
 try:
-    import janitor.xarray
+    import janitor.xarray  # noqa: F401
 except ImportError:
     pass
 
 from .functions import *  # noqa: F403, F401
-from .math import *
+from .math import *  # noqa: F403, F401
 from .ml import get_features_targets as _get_features_targets
 from .utils import refactored_function
 
@@ -19,4 +19,4 @@ def get_features_targets(*args, **kwargs):
     return _get_features_targets(*args, **kwargs)
 
 
-__version__ = "0.20.3"
+__version__ = "0.20.5"