+++ janitor/xarray/__init__.py
@@ -1,1 +1,1 @@
-from .functions import *
+from .functions import *  # noqa: F403, F401