+++ tests/math/test_log.py
@@ -13,12 +13,12 @@ def test_log():
     with pytest.warns(RuntimeWarning):
         out = s.log(error="warn")
 
-    assert out[s <= 0].isnull().all()
+    assert out[s <= 0].isna().all()
     assert (out.index == s.index).all()
     assert (out[s > 0] == np.log(np.array([1, 2, 3]))).all()
 
     out = s.log(error="ignore")
 
-    assert out[s <= 0].isnull().all()
+    assert out[s <= 0].isna().all()
     assert (out.index == s.index).all()
     assert (out[s > 0] == np.log(np.array([1, 2, 3]))).all()