+++ tests/math/test_probit.py
@@ -17,14 +17,14 @@ def test_probit():
     with pytest.warns(RuntimeWarning):
         out = s.probit(error="warn")
 
-    assert out[inside].notnull().all()
+    assert out[inside].notna().all()
     assert (out[inside] == ans).all()
     assert (out.index == s.index).all()
-    assert out[~inside].isnull().all()
+    assert out[~inside].isna().all()
 
     out = s.probit(error="ignore")
 
-    assert out[inside].notnull().all()
+    assert out[inside].notna().all()
     assert (out[inside] == ans).all()
     assert (out.index == s.index).all()
-    assert out[~inside].isnull().all()
+    assert out[~inside].isna().all()