+++ CHANGELOG.rst
@@ -2,6 +2,12 @@ new version (on deck)
 =====================
 
 
+v0.20.5
+=======
+- [INF]: Replace ``pycodestyle`` with ``flake8`` in order to add ``pandas-vet`` linter @hectormz
+- [ENH] ``select_columns()`` now raises ``NameError`` if column label in
+  ``search_columns_labels`` is missing from ``DataFrame`` columns. @smu095
+
 v0.20.1
 =======
 - [DOC]: Added an example for groupby_agg in general functions @samukweku